# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Frida iOS Analyzer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface ConnectRequest {
  /** Frida server host (IP or hostname) */
  host: string;
  /** Frida server port (default 27042) */
  port: number;
}

export interface ConnectionStatus {
  connected: boolean;
  /** @nullable */
  host: string | null;
  /** @nullable */
  port: number | null;
  /** @nullable */
  deviceType: string | null;
  /** @nullable */
  deviceName: string | null;
}

export interface Process {
  pid: number;
  name: string;
  /** @nullable */
  identifier: string | null;
  /** @nullable */
  smallIcon: string | null;
  /** @nullable */
  largeIcon: string | null;
}

export interface Application {
  identifier: string;
  name: string;
  /** @nullable */
  pid: number | null;
  /** @nullable */
  smallIcon: string | null;
  /** @nullable */
  largeIcon: string | null;
}

/**
 * How to identify the target process
 */
export type CreateSessionRequestTargetType =
  (typeof CreateSessionRequestTargetType)[keyof typeof CreateSessionRequestTargetType];

export const CreateSessionRequestTargetType = {
  pid: "pid",
  name: "name",
  identifier: "identifier",
} as const;

export interface CreateSessionRequest {
  /** Process name, bundle ID, or PID */
  target: string;
  /** How to identify the target process */
  targetType: CreateSessionRequestTargetType;
}

export type SessionStatus = (typeof SessionStatus)[keyof typeof SessionStatus];

export const SessionStatus = {
  active: "active",
  detached: "detached",
} as const;

export interface Session {
  id: string;
  target: string;
  targetType: string;
  pid: number;
  processName: string;
  status: SessionStatus;
  createdAt: string;
}

export interface ClassListResponse {
  classes: string[];
  total: number;
  filtered: number;
}

export interface MethodListResponse {
  className: string;
  instanceMethods: string[];
  classMethods: string[];
}

export interface ExecuteScriptRequest {
  /** Frida JavaScript code to execute */
  code: string;
  /** Timeout in milliseconds (default 10000) */
  timeout?: number;
}

export interface ScriptResult {
  output: string[];
  /** @nullable */
  error: string | null;
  duration: number;
}

export type CreateHookRequestMethodType =
  (typeof CreateHookRequestMethodType)[keyof typeof CreateHookRequestMethodType];

export const CreateHookRequestMethodType = {
  instance: "instance",
  class: "class",
} as const;

export interface CreateHookRequest {
  className: string;
  methodName: string;
  methodType: CreateHookRequestMethodType;
  logArgs?: boolean;
  logReturn?: boolean;
}

export interface Hook {
  id: string;
  className: string;
  methodName: string;
  methodType: string;
  logArgs: boolean;
  logReturn: boolean;
  callCount: number;
  createdAt: string;
}

export type ListClassesParams = {
  /**
   * Optional filter string to search class names
   */
  filter?: string;
};