# lib/api-zod/src/generated/types/classListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Frida iOS Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface ClassListResponse {
  classes: string[];
  total: number;
  filtered: number;
}