# lib/api-zod/src/generated/types/connectRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Frida iOS Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface ConnectRequest {
  /** Frida server host (IP or hostname) */
  host: string;
  /** Frida server port (default 27042) */
  port: number;
}