# lib/api-zod/src/generated/types/connectionStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Frida iOS Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface ConnectionStatus {
  connected: boolean;
  /** @nullable */
  host: string | null;
  /** @nullable */
  port: number | null;
  /** @nullable */
  deviceType: string | null;
  /** @nullable */
  deviceName: string | null;
}