# lib/api-zod/src/generated/types/createHookRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Frida iOS Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateHookRequestMethodType } from "./createHookRequestMethodType";

export interface CreateHookRequest {
  className: string;
  methodName: string;
  methodType: CreateHookRequestMethodType;
  logArgs?: boolean;
  logReturn?: boolean;
}