# lib/api-zod/src/generated/types/createSessionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Frida iOS Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateSessionRequestTargetType } from "./createSessionRequestTargetType";

export interface CreateSessionRequest {
  /** Process name, bundle ID, or PID */
  target: string;
  /** How to identify the target process */
  targetType: CreateSessionRequestTargetType;
}