# lib/api-zod/src/generated/types/executeScriptRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Frida iOS Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface ExecuteScriptRequest {
  /** Frida JavaScript code to execute */
  code: string;
  /** Timeout in milliseconds (default 10000) */
  timeout?: number;
}