# lib/api-zod/src/generated/types/hook.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Frida iOS Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface Hook {
  id: string;
  className: string;
  methodName: string;
  methodType: string;
  logArgs: boolean;
  logReturn: boolean;
  callCount: number;
  createdAt: string;
}