# lib/api-zod/src/generated/types/methodListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Frida iOS Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface MethodListResponse {
  className: string;
  instanceMethods: string[];
  classMethods: string[];
}