# lib/api-zod/src/generated/types/process.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Frida iOS Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface Process {
  pid: number;
  name: string;
  /** @nullable */
  identifier: string | null;
  /** @nullable */
  smallIcon: string | null;
  /** @nullable */
  largeIcon: string | null;
}