# lib/api-zod/src/generated/types/scriptResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Frida iOS Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface ScriptResult {
  output: string[];
  /** @nullable */
  error: string | null;
  duration: number;
}