# lib/api-zod/src/generated/types/session.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Frida iOS Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import type { SessionStatus } from "./sessionStatus";

export interface Session {
  id: string;
  target: string;
  targetType: string;
  pid: number;
  processName: string;
  status: SessionStatus;
  createdAt: string;
}